LF-5272 Rename CustomRevenueRadios to CustomRevenueEntityTypeSelector
